выстроена двухуровненная архитектура
