--- addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/GroupCreationTests.java
@@ -1,19 +1,0 @@
-package ru.stqa.pft.addressbook;
-
-
-import org.openqa.selenium.By;
-import org.testng.annotations.Test;
-
-
-public class GroupCreationTests extends TestBase {
-
-    @Test
-    public void testGroupCreation() {
-        gotoGroupPage();
-        initGroupCreation();
-        fillGroupForm(new GroupData("test2", "test3", "test4"));
-        submitGroupCreation();
-        returnToGroupPage();
-    }
-
-}
