--- addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/GroupDeletionTests.java
@@ -1,24 +1,0 @@
-package ru.stqa.pft.addressbook;
-
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.Test;
-
-public class GroupDeletionTests extends TestBase {
-
-
-    
-    @Test
-    public void testGroupDeletionTests() {
-        gotoGroupPage();
-        selectGroup();
-        deleteGroup();
-        returnToGroupPage();
-    }
-
-    @AfterMethod
-    public void tearDown() {
-        wd.quit();
-    }
-
-
-}
