+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/ApplicationManager.java
@@ -1,0 +1,45 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.NoAlertPresentException;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
+
+import java.util.concurrent.TimeUnit;
+
+public class ApplicationManager {
+  FirefoxDriver wd;
+  private SessionHelper sessionHelper;
+  private NaviationHelper naviationHelper;
+  private GroupHelper groupHelper;
+
+  public static boolean isAlertPresent(FirefoxDriver wd) {
+    try {
+      wd.switchTo().alert();
+      return true;
+    } catch (NoAlertPresentException e) {
+      return false;
+    }
+  }
+
+  public void init() {
+    wd = new FirefoxDriver(new FirefoxOptions().setLegacy(true));
+    wd.manage().timeouts().implicitlyWait(60, TimeUnit.SECONDS);
+    wd.get("http://localhost/addressbook/group.php");
+    naviationHelper = new NaviationHelper(wd);
+    sessionHelper.login("admin", "secret");
+    sessionHelper = new SessionHelper(wd);
+  }
+
+
+  public void stop() {
+    wd.quit();
+  }
+
+  public GroupHelper getGroupHelper() {
+    return groupHelper;
+  }
+
+  public NaviationHelper getNaviationHelper() {
+    return naviationHelper;
+  }
+}
