+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/GroupHelper.java
@@ -1,0 +1,44 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import ru.stqa.pft.addressbook.model.GroupData;
+
+public class GroupHelper {
+  public FirefoxDriver wd;
+
+  public GroupHelper(FirefoxDriver wd) {
+  }
+
+  public void returnToGroupPage() {
+      wd.findElement(By.linkText("group page")).click();
+  }
+
+  public void submitGroupCreation() {
+      wd.findElement(By.name("submit")).click();
+  }
+
+  public void fillGroupForm(GroupData groupData) {
+      wd.findElement(By.name("group_name")).click();
+      wd.findElement(By.name("group_name")).clear();
+      wd.findElement(By.name("group_name")).sendKeys(groupData.getHeader());
+      wd.findElement(By.name("group_header")).click();
+      wd.findElement(By.name("group_header")).clear();
+      wd.findElement(By.name("group_header")).sendKeys(groupData.getBody());
+      wd.findElement(By.name("group_footer")).click();
+      wd.findElement(By.name("group_footer")).clear();
+      wd.findElement(By.name("group_footer")).sendKeys(groupData.getFooter());
+  }
+
+  public void initGroupCreation() {
+      wd.findElement(By.name("new")).click();
+  }
+
+  public void deleteGroup() {
+      wd.findElement(By.name("delete")).click();
+  }
+
+  public void selectGroup() {
+      wd.findElement(By.name("selected[]")).click();
+  }
+}
