+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/NaviationHelper.java
@@ -1,0 +1,16 @@
+package ru.stqa.pft.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+public class NaviationHelper {
+ private  FirefoxDriver wd;
+
+  public NaviationHelper(FirefoxDriver wd) {
+this.wd = wd;
+  }
+
+  public void gotoGroupPage() {
+    wd.findElement(By.linkText("groups")).click();
+  }
+}
