+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/model/GroupData.java
@@ -1,4 +1,4 @@
-package ru.stqa.pft.addressbook;
+package ru.stqa.pft.addressbook.model;
 
 public class GroupData {
   private final String header;
