+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/tests/GroupCreationTests.java
@@ -1,0 +1,19 @@
+package ru.stqa.pft.addressbook.tests;
+
+
+import org.testng.annotations.Test;
+import ru.stqa.pft.addressbook.model.GroupData;
+
+
+public class GroupCreationTests extends TestBase {
+
+    @Test
+    public void testGroupCreation() {
+        app.getNaviationHelper().gotoGroupPage();
+        app.getGroupHelper().initGroupCreation();
+        app.getGroupHelper().fillGroupForm(new GroupData("test2", "test3", "test4"));
+        app.getGroupHelper().submitGroupCreation();
+        app.getGroupHelper().returnToGroupPage();
+    }
+
+}
