+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/tests/TestBase.java
@@ -1,0 +1,21 @@
+package ru.stqa.pft.addressbook.tests;
+
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+import ru.stqa.pft.addressbook.appmanager.ApplicationManager;
+
+public class TestBase {
+
+  protected final ApplicationManager app = new ApplicationManager();
+
+  @BeforeMethod
+  public void setUp() throws Exception {
+    app.init();
+  }
+
+  @AfterMethod
+  public void tearDown() {
+    app.stop();
+  }
+
+}
